dev/requirements.txt
flask>=3.0
requests>=2.31
pillow>=10.0
python-dotenv>=1.0
